+++ mashkin.andrew/I2/maxCountOfEqualElements.h
@@ -2,5 +2,5 @@
 #define MAXCOUNTOFEQUALELEMENTS_H
 #include <cstddef>
 
-void countMaxQuantityOfEqualElements(const int* arr, const int size, size_t& count, size_t& maxCount);
+void countMaxQuantityOfEqualElements(const int* arr, const size_t size, size_t& count, size_t& maxCount);
 #endif
