+++ mashkin.andrew/I2/samePositiveElements.cpp
@@ -1,6 +1,6 @@
 #include "samePositiveElements.h"
 
-bool isThereTheSamePositiveElements(const int* arr, const int size, bool& ans)
+bool isThereTheSamePositiveElements(const int* arr, const size_t size, bool& ans)
 {
   for (int i = 0; i < size - 1; ++i)
   {
